Changing cfg organisation to enable multiple functions
The grammar is changed to allow multiple functions
instead of having one cfg the CToIRVisitor has a list of them
(one for each function) along with a pointer to the current cfg.
The cfgs are not instanciated in the CToIRVisitor constructor,
but instead are instanciated when visiting a function.
We add a method in CToIRVisitor.h to add a new CFG and update the
current cfg pointer when visiting a new function.